requirements.txt
asgiref==3.5.2
Django==3.0.14
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
pytz==2022.6
sqlparse==0.4.3
